requirements.txt
asgiref==3.6.0
boto3==1.26.54
botocore==1.29.54
Django==4.1.5
django-etc==1.4.0
django-hitcount==1.3.5
django-storages==1.13.2
django_debug_toolbar==3.8.1
environs==9.5.0
jmespath==1.0.1
marshmallow==3.19.0
mysqlclient==2.1.1
packaging==23.0
Pillow==9.4.0
python-dateutil==2.8.2
python-dotenv==0.21.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
shortuuid